# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SMART-ACADEMY API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  /** @nullable */
  excerpt?: string | null;
  content: string;
  /** @nullable */
  thumbnail?: string | null;
  authorId: number;
  authorName: string;
  /** @nullable */
  authorAvatar?: string | null;
  category: string;
  /** @nullable */
  tags?: string | null;
  viewCount: number;
  likeCount: number;
  readMinutes: number;
  publishedAt: string;
}