# lib/api-zod/src/generated/types/communityPostInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SMART-ACADEMY API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CommunityPostInput {
  /** @minLength 1 */
  content: string;
  authorId: number;
  authorName: string;
  authorAvatar?: string;
  authorRole: string;
  topic?: string;
}