# lib/api-zod/src/generated/types/courseUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SMART-ACADEMY API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CourseUpdate {
  title?: string;
  description?: string;
  shortDescription?: string;
  thumbnail?: string;
  price?: number;
  isFeatured?: boolean;
  level?: string;
  tags?: string;
}