# lib/api-zod/src/generated/types/enrollment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SMART-ACADEMY API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Enrollment {
  id: number;
  userId: number;
  courseId: number;
  /** @nullable */
  courseTitle?: string | null;
  /** @nullable */
  courseThumbnail?: string | null;
  progress: number;
  /** @nullable */
  completedAt?: string | null;
  enrolledAt: string;
  /** @nullable */
  certificateUrl?: string | null;
}